vktrace: Rename vkparser to vktraceparser
This change renames vkparser in previous commit to vktraceparser.
